+++ tests/test_osint.py
@@ -3,6 +3,7 @@
 
 """
 import os
+import sys
 import io
 from random import randbytes
 import logging
@@ -10,7 +11,9 @@ import logging
 import pytest
 
 from sphinxcontrib import osint
-from sphinxcontrib.osint.osintlib import collect, osint_plugins
+from sphinxcontrib.osint.plugins import collect_plugins
+
+sys.path.append(os.path.abspath(".."))
 
 cats = {
     'test1' : {
@@ -224,10 +227,36 @@ ident_ident1 -> event_event1 [label="link1"];
 
 def test_plugins(caplog):
     caplog.set_level(logging.DEBUG, logger="osint")
-# ~ def test_osint_plugins():
-    plgs = collect()
-    # ~ print(plgs)
+    plgs = collect_plugins()
     print(plgs)
     print(plgs['source'])
     print(plgs['source'][0].order)
+    # ~ assert False
+
+def test_bsky_regexp(caplog):
+    caplog.set_level(logging.DEBUG, logger="osint")
+    from sphinxcontrib.osint.plugins import bskylib
+    import private_conf
+    handle, post = bskylib.BSkyInterface.post2atp('https://bsky.app/profile/wsj.com/post/3luuxgt5sye2y')
+    assert handle == 'wsj.com'
+    assert post == '3luuxgt5sye2y'
+    handle, post = bskylib.BSkyInterface.post2atp('https://bsky.app/badpost')
+    assert handle is None
+    assert post is None
+    handle = bskylib.BSkyInterface.profile2atp('https://bsky.app/profile/wsj.com/post/3luuxgt5sye2y')
+    assert handle == 'wsj.com'
+    handle, post = bskylib.BSkyInterface.post2atp('https://bsky.app/badpost')
+    assert handle is None
+
+def test_bsky_get(caplog):
+    caplog.set_level(logging.DEBUG, logger="osint")
+    from sphinxcontrib.osint.plugins import bskylib
+    import private_conf
+    import json
+    bsk_get = bskylib.OSIntBSkyGet('name', 'label', quest='fake',
+        user=private_conf.osint_bsky_user, apikey=private_conf.osint_bsky_apikey,
+        url='https://bsky.app/profile/wsj.com/post/3luuxgt5sye2y')
+    resp = bsk_get.get()
+    print(json.dumps(resp.__dict__), indent=4)
     assert False
+